+++ src/tbp/monty/frameworks/models/mixins/no_reset_evidence.py
@@ -11,7 +11,9 @@ from typing import Any, Dict
 
 from scipy.spatial.transform import Rotation
 
-from tbp.monty.frameworks.models.evidence_matching import EvidenceGraphLM
+from tbp.monty.frameworks.models.evidence_matching.evidence_matching import (
+    EvidenceGraphLM,
+)
 from tbp.monty.frameworks.utils.logging_utils import compute_pose_error
